v2.2.8 Alpha, 增加参数配置页，多项优化

+++ src/core/interface.py
@@ -133,7 +133,7 @@ class OCRService(ABC):
     #     pass
 
     @abstractmethod
-    def wait_text(self, targets: str | list[str], timeout: int = 3,
+    def wait_text(self, targets: str | list[str], timeout: float = 3.0,
                   position: Position | DynamicPosition | None = None, wait_time: float = 0.1) -> TextPosition | None:
         pass
